+++ src/Language/Elm/Definition.hs
@@ -2,6 +2,9 @@ module Language.Elm.Definition where
 
 import Data.Void
 
+import Bound (Scope)
+import qualified Bound
+
 import Language.Elm.Expression (Expression)
 import qualified Language.Elm.Expression as Expression
 import qualified Language.Elm.Name as Name
@@ -9,15 +12,15 @@ import Language.Elm.Type (Type)
 import qualified Language.Elm.Type as Type
 
 data Definition
-  = Constant !Name.Qualified (Type Void) (Expression Void)
-  | Type !Name.Qualified [(Name.Constructor, [Type Void])]
-  | Alias !Name.Qualified (Type Void)
+  = Constant !Name.Qualified !Int (Scope Int Type Void) (Expression Void)
+  | Type !Name.Qualified !Int [(Name.Constructor, [Scope Int Type Void])]
+  | Alias !Name.Qualified !Int (Scope Int Type Void)
   deriving (Eq, Ord, Show)
 
 name :: Definition -> Name.Qualified
-name (Constant n _ _) = n
-name (Type n _) = n
-name (Alias n _) = n
+name (Constant n _ _ _) = n
+name (Type n _ _) = n
+name (Alias n _ _) = n
 
 foldMapGlobals
   :: Monoid m
@@ -26,15 +29,15 @@ foldMapGlobals
   -> m
 foldMapGlobals f def =
   case def of
-    Constant qname type_ expr ->
+    Constant qname _ type_ expr ->
       f qname <>
-      Type.foldMapGlobals f type_ <>
+      Type.foldMapGlobals f (Bound.fromScope type_) <>
       Expression.foldMapGlobals f expr
 
-    Type qname constrs ->
+    Type qname _ constrs ->
       f qname <>
-      foldMap (foldMap (foldMap (Type.foldMapGlobals f))) constrs
+      foldMap (foldMap (foldMap (Type.foldMapGlobals f . Bound.fromScope))) constrs
 
-    Alias qname type_ ->
+    Alias qname _ type_ ->
       f qname <>
-      Type.foldMapGlobals f type_
+      Type.foldMapGlobals f (Bound.fromScope type_)